requirements.txt
contourpy==1.3.2
cycler==0.12.1
Cython==0.29.37
fonttools==4.61.1
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scikit-surprise==1.1.4
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
